feat: Forge related changes, renamed compat to integration, and edited buildscript
