+++ build.gradle
@@ -109,6 +109,9 @@ dependencies {
     compileOnly "top.theillusivec4.caelus:caelus-forge:${project.caelus_version}:api"
     compileOnly "top.theillusivec4.caelus:caelus-forge:${project.caelus_version}"
 
+    compileOnly "mezz.jei:jei-${project.mc_version}-common-api:${project.jei_version}"
+    compileOnly "mezz.jei:jei-${project.mc_version}-forge-api:${project.jei_version}"
+
 //    jarJar fg.deobf("com.gildedgames.nitrogen:nitrogen:${project.nitrogen_version}") {
 //        jarJar.ranged(it, "[${project.nitrogen_version},)")
 //        jarJar.pin(it, "${project.nitrogen_version}")
@@ -127,10 +130,6 @@ dependencies {
     //runtimeOnly fg.deobf("io.github.ultimateboomer:smoothboot-forge:${project.smoothboot_version}")
 
     //compileOnly fileTree(dir: 'libs', include: '*.jar')
-
-    compileOnly(fg.deobf("mezz.jei:jei-${mc_version}-common-api:${jei_version}"))
-    compileOnly(fg.deobf("mezz.jei:jei-${mc_version}-forge-api:${jei_version}"))
-    runtimeOnly(fg.deobf("mezz.jei:jei-${mc_version}-forge:${jei_version}"))
 }
 
 repositories {
