+++ src/main/java/com/gildedgames/aether/integration/jei/categories/fuel/AetherFuelCategory.java
@@ -1,4 +1,4 @@
-package com.gildedgames.aether.compat.jei.categories.fuel;
+package com.gildedgames.aether.integration.jei.categories.fuel;
 
 import com.gildedgames.aether.Aether;
 import com.gildedgames.aether.block.AetherBlocks;
